extract Scrollbars
